% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dicionario_radarsp.R
\name{dicionario_radarsp}
\alias{dicionario_radarsp}
\title{Download São Paulo speed camera data}
\usage{
dicionario_radarsp(as_data_frame = TRUE, show_progress = TRUE, cache = TRUE)
}
\arguments{
\item{as_data_frame}{Logical. Whether to save a dataframe or an Arrow object. Defaults to \code{TRUE}.}

\item{show_progress}{Logical. Whether to show progress bar. Defaults to \code{TRUE}.}

\item{cache}{Logical. Whether to show progress bar. Defaults to \code{TRUE}.}
}
\value{
An arrow \code{Dataset} or a \code{"data.frame"} object.
}
\description{
Download data dictionary for package "radarsp"
}
\examples{
\dontshow{if (identical(tolower(Sys.getenv("NOT_CRAN")), "true")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# return data as arrow Dataset
df <- dicionario_radarsp(as_data_frame = TRUE)
 
\dontshow{\}) # examplesIf}
}
\seealso{
Other Microdata: 
\code{\link{read_15min}()},
\code{\link{read_day}()},
\code{\link{read_hour}()}
}
\concept{Microdata}
