% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_file}
\alias{download_file}
\title{Download file from url}
\usage{
download_file(file_url = file_url, showProgress = show_progress, cache = cache)
}
\arguments{
\item{file_url}{String. A url.}

\item{showProgress}{Logical.}

\item{cache}{Logical.}
}
\value{
A string to the address of the file in a tempdir
}
\description{
Download file from url
}
\keyword{internal}
