% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radarsp.R
\docType{package}
\name{radarsp}
\alias{radarsp}
\alias{radarsp-package}
\title{radarsp: Baixe dados dos radares de trânsito da cidade de São Paulo, Brasil.}
\description{
Baixar dados dos radares de trânsito da cidade de São Paulo, Brasil.
}
\section{Usage}{

Por favor verifique nossas vinhetas e documentação em
\href{https://github.com/ciclocidade/radares_sp/}{website}.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ciclocidade/radares_sp}
  \item Report bugs at \url{https://github.com/ciclocidade/radares_sp/issues}
}

}
\author{
\strong{Maintainer}: Tainá S. Pacheco \email{pacheco.taina@gmail.com}

Authors:
\itemize{
  \item Flávio Soares \email{flavio@ciclocidade.org.br}
}

Other contributors:
\itemize{
  \item Rafael H. M. Pereira \email{rafa.pereira.br@gmail.com} [contributor]
  \item Ciclocidade - Associação dos ciclistas urbanos da cidade de São Paulo [copyright holder, funder]
}

}
\keyword{internal}
