% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radarsp_cache.R
\name{radarsp_cache}
\alias{radarsp_cache}
\title{Manage cached files from the radarsp package}
\usage{
radarsp_cache(list_files = TRUE, delete_file = NULL)
}
\arguments{
\item{list_files}{Logical. Whether to print a message with the address of all
radarsp data sets cached locally. Defaults to \code{TRUE}.}

\item{delete_file}{String. The file name (basename) of a radarsp data set
cached locally that should be deleted. Defaults to \code{NULL}, so that no
file is deleted. If \code{delete_file = "all"}, then all cached radarsp
files are deleted.}
}
\value{
A message indicating which file exist and/or which ones have been
deleted from local cache directory.
}
\description{
Manage cached files from the radarsp package
}
\examples{
\dontshow{if (identical(tolower(Sys.getenv("NOT_CRAN")), "true")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# list all files cached
radarsp_cache(list_files = TRUE)

# delete particular file
radarsp_cache(delete_file = '20190101_24')
\dontshow{\}) # examplesIf}
}
\seealso{
Other Cache data: 
\code{\link{set_radarsp_cache_dir}()}
}
\concept{Cache data}
