% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_day.R
\name{read_day}
\alias{read_day}
\title{Download São Paulo speed camera data}
\usage{
read_day(
  start = "2019/01/01",
  end = NULL,
  id_to_filter = NULL,
  as_data_frame = FALSE,
  show_progress = TRUE,
  cache = TRUE
)
}
\arguments{
\item{start}{String. Data to start the search. Format 'YYYY/MM/DD'. Defaults to \code{"2019/01/01"}.}

\item{end}{String. Data to start the search. Format 'YYYY/MM/DD'. Defaults to \code{NULL}, so that end data equals to start date.}

\item{id_to_filter}{String. Ids of speed cameras to filter for. Must be a 4 character string. Defaults to \code{NULL}, so all data is provided.}

\item{as_data_frame}{Logical. Whether to save a dataframe or an Arrow object. Defaults to \code{TRUE}.}

\item{show_progress}{Logical. Whether to show progress bar. Defaults to \code{TRUE}.}

\item{cache}{Logical. Whether to show progress bar. Defaults to \code{TRUE}.}
}
\value{
An arrow \code{Dataset} or a \code{"data.frame"} object.
}
\description{
Download speed camera records grouped by location and day.
}
\examples{
\dontshow{if (identical(tolower(Sys.getenv("NOT_CRAN")), "true")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# return data as arrow Dataset
df <- read_hour(start = "2019/01/10",
                as_data_frame = TRUE)
                
\dontshow{\}) # examplesIf}
}
\seealso{
Other Microdata: 
\code{\link{dicionario_radarsp}()},
\code{\link{read_15min}()},
\code{\link{read_hour}()}
}
\concept{Microdata}
